+++ upgrade.js
@@ -1,6 +1,7 @@
 export const upgradeScripts = [
     function removeOldInvertFeedbackScripts(context, props) {
         var result = {};
+        result.updatedFeedbacks = [];
         for (const feedback of props.feedbacks) {
             if (feedback.options['invertFeedback'] !== undefined) {
                 feedback.isInverted = !!feedback.options['invertFeedback']
